updated: change product id type from int to Long for synchronize data type with Inventory

+++ src/main/java/com/learningmat/ecommerce/module/product/ProductController.java
@@ -15,7 +15,7 @@ import java.util.List;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/products")
-@Tag(name="Product Management", description = "API for product including get item list and management")
+@Tag(name = "Product Management", description = "API for product including get item list and management")
 public class ProductController {
     private final ProductService productService;
 
@@ -23,7 +23,8 @@ public class ProductController {
     @ResponseStatus(HttpStatus.CREATED)
     @PreAuthorize("hasRole('STAFF')")
     @Operation(summary = "create a new product", description = "for create a new product, must be staff role or higher to use this")
-    public ApiResponse<Product> createProduct(@RequestBody @Valid ProductRequest productRequest) {
+    public ApiResponse<Product> createProduct(
+            @RequestBody @Valid ProductRequest productRequest) {
         Product prod = productService.createProduct(productRequest);
         return ApiResponse.<Product>builder()
                 .result(prod)
@@ -45,7 +46,8 @@ public class ProductController {
     @GetMapping("/{productId}")
     @PreAuthorize("hasRole('USER')")
     @Operation(summary = "Get product by its id")
-    public ApiResponse<Product> getProductById(@PathVariable int productId) {
+    public ApiResponse<Product> getProductById(
+            @PathVariable Long productId) {
         return ApiResponse.<Product>builder()
                 .result(productService.getProductById(productId))
                 .build();
@@ -54,7 +56,8 @@ public class ProductController {
     @PutMapping("/{productId}")
     @PreAuthorize("hasRole('STAFF')")
     @Operation(method = "???", summary = "update product information", description = "for update product and must have higher role to use this")
-    public ApiResponse<Product> updateProduct(@PathVariable int productId,
+    public ApiResponse<Product> updateProduct(
+            @PathVariable Long productId,
             @RequestBody @Valid ProductRequest productRequest) {
         return ApiResponse.<Product>builder()
                 .result(productService.updateProduct(productId, productRequest))
@@ -65,7 +68,8 @@ public class ProductController {
     @DeleteMapping("/{productId}")
     @PreAuthorize("hasRole('STAFF')")
     @Operation(summary = "delete a product")
-    public ApiResponse<String> deleteProduct(@PathVariable int productId) {
+    public ApiResponse<String> deleteProduct(
+            @PathVariable Long productId) {
         productService.deleteProduct(productId);
         return ApiResponse.<String>builder()
                 .message("Product delete successfully")
